+++ obj/src/persistence/MongoDbIndex.d.ts
@@ -1,0 +1,13 @@
+/**
+ * Index definition for mondodb
+ */
+export declare class MongoDbIndex {
+    /**
+     * Index keys (fields)
+     */
+    keys: any;
+    /**
+     * Index options
+     */
+    options?: any;
+}