+++ obj/src/persistence/MongoDbIndex.js
@@ -1,0 +1,9 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+/**
+ * Index definition for mondodb
+ */
+class MongoDbIndex {
+}
+exports.MongoDbIndex = MongoDbIndex;
+//# sourceMappingURL=MongoDbIndex.js.map