+++ src/persistence/MongoDbPersistence.ts
@@ -1,4 +1,8 @@
 /** @module persistence */
+
+let _ = require('lodash');
+let async = require('async');
+
 import { IReferenceable } from 'pip-services3-commons-node';
 import { IUnreferenceable } from 'pip-services3-commons-node';
 import { IReferences } from 'pip-services3-commons-node';
@@ -13,6 +17,7 @@ import { CompositeLogger } from 'pip-services3-components-node';
 
 import { MongoDbConnectionResolver } from '../connect/MongoDbConnectionResolver';
 import { MongoDbConnection } from './MongoDbConnection';
+import { MongoDbIndex } from './MongoDbIndex';
 
 /**
  * Abstract persistence component that stores data in MongoDB using plain driver.
@@ -111,6 +116,7 @@ export class MongoDbPersistence implements IReferenceable, IUnreferenceable, ICo
     private _references: IReferences;
     private _opened: boolean;
     private _localConnection: boolean;
+    private _indexes: MongoDbIndex[] = [];
 
     /**
      * The dependency resolver.
@@ -210,6 +216,21 @@ export class MongoDbPersistence implements IReferenceable, IUnreferenceable, ICo
         return connection;
     }
 
+    /**
+     * Adds index definition to create it on opening
+     * @param keys index keys (fields)
+     * @param options index options
+     */
+    protected ensureIndex(keys: any, options?: any): void {
+        if (keys == null) return;
+        this._indexes.push(
+            <MongoDbIndex> {
+                keys: keys,
+                options: options
+            }
+        );
+    }
+
     /** 
      * Converts object value from internal to public format.
      * 
@@ -291,13 +312,33 @@ export class MongoDbPersistence implements IReferenceable, IUnreferenceable, ICo
                         this._db = null;
                         this._client == null;
                         err = new ConnectionException(correlationId, "CONNECT_FAILED", "Connection to mongodb failed").withCause(err);
-                    } else {
-                        this._opened = true;
-                        this._collection = collection;
-                        this._logger.debug(correlationId, "Connected to mongodb database %s, collection %s", this._databaseName, this._collectionName);
+                        if (callback) callback(err);
+                        return;
                     }
 
-                    if (callback) callback(err);
+                    // Recreate indexes
+                    async.each(this._indexes, (index, callback) => {
+                        collection.createIndex(index.keys, index.options, (err) => {
+                            if (err == null) {
+                                let options = index.options || {};
+                                let indexName = options.name || _.keys(index.keys).join(',');
+                                this._logger.debug(correlationId, "Created index %s for collection %s", indexName, this._collectionName);
+                            }
+                            callback(err);
+                        });
+                    }, (err) => {
+                        if (err) {
+                            this._db = null;
+                            this._client == null;
+                            err = new ConnectionException(correlationId, "CONNECT_FAILED", "Connection to mongodb failed").withCause(err);    
+                        } else {
+                            this._opened = true;
+                            this._collection = collection;        
+                            this._logger.debug(correlationId, "Connected to mongodb database %s, collection %s", this._databaseName, this._collectionName);                        
+                        }
+
+                        if (callback) callback(err);
+                    });
                 });
             }
         };